+++ src/hooks/accountingAccounts/useAccountingAccountsByClient.ts
@@ -1,7 +1,14 @@
-import { IGetAccountingAccountsByClientDTO } from 'src/modules/clients/dtos/IGetAccountingAccountsByClientDTO'
+// DTOs
+import { IGetAccountingAccountsByClientDTO } from 'src/modules/accountingAccounts/dtos/IGetAccountingAccountsByClientDTO'
+import { IGetAccountingAccountsByClientResponseDTO } from 'src/modules/accountingAccounts/dtos/IGetAccountingAccountsByClientResponseDTO'
+
+// React Query
 import { useQuery, UseQueryOptions } from 'react-query'
-import { IGetAccountingAccountsByClientResponseDTO } from 'src/modules/clients/dtos/IGetAccountingAccountsByClientResponseDTO'
-import { clientsController } from 'src/modules/clients'
+
+// Controllers
+import { accountingAccountsController } from 'src/modules/accountingAccounts'
+
+// Errors
 import { AppError } from 'src/shared/errors/AppError'
 
 export function useAccountingAccountsByClient(
@@ -15,7 +22,7 @@ export function useAccountingAccountsByClient(
 ) {
   return useQuery(
     ['accounting-accounts-by-client', data],
-    () => clientsController.getAccountingAccountsByClient(data),
+    () => accountingAccountsController.getAccountingAccountsByClient(data),
     options
   )
 }
