+++ src/modules/accountingAccounts/dtos/IDeleteAccountingAccountDTO.ts
