+++ src/modules/accountingAccounts/dtos/IGetAccountingAccountsByClientResponseDTO.ts
