+++ src/modules/accountingAccounts/useCases/CreateAccountingAccountUseCase.ts
@@ -1,13 +1,16 @@
-import { IClientsRepository } from '../repositories/IClientsRepository'
+// Interfaces
+import { IAccountingAccountsRepository } from '../repositories/IAccountingAccountsRepository'
 import { ICreateAccountingAccountDTO } from '../dtos/ICreateAccountingAccountDTO'
+
+// Providers
 import { errorProvider } from 'src/shared/providers'
 
 export class CreateAccountingAccountUseCase {
-  constructor(private clientsRepository: IClientsRepository) {}
+  constructor(private accountingAccountsRepository: IAccountingAccountsRepository) {}
 
   async execute(data: ICreateAccountingAccountDTO) {
     try {
-      await this.clientsRepository.createAccountingAccount(data)
+      await this.accountingAccountsRepository.createAccountingAccount(data)
     } catch (e) {
       throw errorProvider.handle(e, {}, 'Erro ao criar conta contábil, tente novamente mais tarde.')
     }
