+++ src/modules/accountingAccounts/useCases/DeleteAccountingAccountUseCase.ts
@@ -1,13 +1,17 @@
-import { IClientsRepository } from '../repositories/IClientsRepository'
+// Interfaces
+import { IAccountingAccountsRepository } from '../repositories/IAccountingAccountsRepository'
 import { IDeleteAccountingAccountDTO } from '../dtos/IDeleteAccountingAccountDTO'
+
+// Providers
 import { errorProvider } from 'src/shared/providers'
 
+
 export class DeleteAccountingAccountUseCase {
-  constructor(private clientsRepository: IClientsRepository) {}
+  constructor(private accountingAccountsRepository: IAccountingAccountsRepository) {}
 
   async execute(data: IDeleteAccountingAccountDTO) {
     try {
-      await this.clientsRepository.deleteAccountingAccount(data)
+      await this.accountingAccountsRepository.deleteAccountingAccount(data)
     } catch (e) {
       throw errorProvider.handle(e, {}, 'Erro ao deletar conta contábil, tente novamente mais tarde.')
     }
