+++ src/modules/accountingAccounts/useCases/GetAccountingAccountsByClientUseCase.ts
@@ -1,13 +1,16 @@
-import { errorProvider } from 'src/shared/providers'
+// Interfaces
 import { IGetAccountingAccountsByClientDTO } from '../dtos/IGetAccountingAccountsByClientDTO'
-import { IClientsRepository } from '../repositories/IClientsRepository'
+import { IAccountingAccountsRepository } from '../repositories/IAccountingAccountsRepository'
+
+// Providers
+import { errorProvider } from 'src/shared/providers'
 
 export class GetAccountingAccountsByClientUseCase {
-  constructor(private clientsRepository: IClientsRepository) {}
+  constructor(private accountingAccountsRepository: IAccountingAccountsRepository) {}
 
   async execute(data: IGetAccountingAccountsByClientDTO) {
     try {
-      return this.clientsRepository.getAccountingAccountsByClient(data)
+      return this.accountingAccountsRepository.getAccountingAccountsByClient(data)
     } catch (e) {
       throw errorProvider.handle(e, {}, 'Erro ao buscar contas contábeis, tente novamente mais tarde.')
     }
